Making posting of only one Inspection parameter mandatory
Providing an option NA each for inspection parameter in case it is not application

Refactored the page into 3 components-CropCycle fields,InspectionParam fields and Adhoc Activity fields
