+++ src/Pages/AddInspectionPage.js
@@ -1,12 +1,8 @@
 import {
-    FormControl,
     FormGroup,
     Grid,
-    InputLabel,
-    MenuItem,
-    Select,
     Snackbar,
-    Typography,Button,TextField, Radio, RadioGroup, FormControlLabel, FormLabel
+    Typography,Button,TextField
   } from '@material-ui/core';
   import { makeStyles } from '@material-ui/core/styles';
   import { Alert } from '@material-ui/lab';
@@ -15,13 +11,13 @@ import {
     MuiPickersUtilsProvider,
   } from '@material-ui/pickers';
   import DateFnsUtils from '@date-io/date-fns';
-  import { getCropCyclesByField } from 'dataclients/CropCyclesClient';
-  import { getFieldsByLocation } from 'dataclients/FieldsClient';
   import { getLocations } from 'dataclients/LocationsClient';
   import { postActivity } from '../dataclients/ProcessesClient'
-  import { getInspectionParams, getAdhocActivities, postInspection } from '../dataclients/InspectionClient';
+  import { getAdhocActivities, postInspection } from '../dataclients/InspectionClient';
   import React, { useEffect, useState } from 'react';
-  
+  import InspectionOptions from '../components/InspectionOptions';
+  import InspectionAdHocActivity from '../components/InspectionAdHocActivity';
+  import CropCycleComponent from '../components/CropCycleComponent';
   const useStyles = makeStyles((theme) => ({
     gridItem: {
       marginLeft: 'auto',
@@ -54,13 +50,9 @@ import {
   
   const AddInspectionPage = () => {
     const classes = useStyles();
-    const [locationCode, setLocationCode] = useState('');
     const [locations, setLocations] = useState([]);
-    const [fieldId, setFieldId] = useState('');
-    const [fields, setFields] = useState([]);
     const [cropCycleId, setCropCycleId] = useState('');
     const [cropCycles, setCropCycles] = useState([]);
-    const [inspectionParams, setInspectionParams] = useState([]);
     const [inspectionVal, setInspectionVal] = useState([]);
     const [selectedInspectionDate, setSelectedInspectionDate] = useState(
         new Date(),
@@ -73,36 +65,12 @@ import {
     const [alertSeverity, setAlertSeverity] = useState('');
     
     const resetForm = () => {
-      setLocationCode('');
-      setFieldId('');
-      setFields([]);
-      initializeInspectionVal();
-      setCropCycles([]);
       setCropCycleId('');
       setSelectedInspectionDate(new Date());
       setShowFollowUpActivity(false);
       setFollowUpActivties([{ code: "", displayStr: "" }]);
       setAdHocActivityList([]);
     }
-    const handleInputChange = (e, index) => {
-        const { name, value } = e.target;
-        const list = [...followUpActivities];
-        list[index].code = value;
-        list[index].displayStr = adHocActivityList.filter((activity) => activity.code === value);
-        setFollowUpActivties(list);
-      };
-    // handle click event of the Remove button
-    const handleRemoveClick = index => {
-        const list = [...followUpActivities];
-        list.splice(index, 1);
-        setFollowUpActivties(list);
-    };
-
-    // handle click event of the Add button
-    const handleAddClick = () => {
-        setFollowUpActivties([...followUpActivities, { code: "", displayStr: "" }]);
-    };
-
   
     const handleAlertClose = (_event, reason) => {
       if (reason === 'clickaway') {
@@ -120,97 +88,7 @@ import {
       setAlertStatus(true);
     };
   
-    useEffect(() => {
-      getLocations()
-        .then(setLocations)
-        .catch((e) => {
-          console.log('Fetching location list failed', e);
-          showAlert('Fetching location list failed', 'error');
-        });
-  
-        getInspectionParams()
-        .then(setInspectionParams)
-        .catch((e) => {
-          console.log('Fetching Inspection Param list failed', e);
-          showAlert('Fetching Inspection Param list failed', 'error');
-        });
-    }, []);
-    
 
-
-    /**
-     * Function to fetch fields for a selected location using API
-     * updates fieldsData if call is successful
-     * shows alert in case call fails
-     * @param {number} selectedLocationCode The code for the selected location
-     */
-    const fetchFieldsForLocation = (selectedLocationCode) => {
-      const displayStr = locations.filter(
-        (obj) => obj.code === selectedLocationCode,
-      )[0].displayStr;
-      getFieldsByLocation(selectedLocationCode)
-        .then(setFields)
-        .catch((e) => {
-          console.log(
-            `Fetching fields for ${selectedLocationCode} - ${displayStr} failed`,
-            e,
-          );
-          showAlert(`Fetching fields for ${displayStr} failed`, 'error');
-        });
-    };
-  
-    /**
-     * Function to fetch CropCycles for a selected field
-     * updates CropCycles if call is successful
-     * shows alert in case call fails
-     * @param {number} selectedFieldId The ID of the selected field
-     */
-    const fetchCropCyclesForField = (selectedFieldId) => {
-      const fieldName = fields.filter((obj) => obj.id === selectedFieldId)[0]
-        .identifier;
-      getCropCyclesByField(selectedFieldId)
-        .then((data) => data.content)
-        .then(formatCropCyclesForDisplay)
-        .then(setCropCycles)
-        .catch((e) => {
-          console.log(
-            `Fetching crop cycles for ${selectedFieldId} - ${fieldName} failed`,
-            e,
-          );
-          showAlert(`Fetching crop cycles for ${fieldName} failed`, 'error');
-        });
-    };
-  
-    /**
-     * Formats the crop cycle objects for display. It addes name as a new
-     * attribute in the object which is formed by using the crop name,
-     * season and year separated by hyphens for better readability
-     * @param {object} cropCycleObjects Objects containing crop cycle details
-     */
-    const formatCropCyclesForDisplay = (cropCycleObjects) =>
-      cropCycleObjects.map((cropCycleObject) => ({
-        ...cropCycleObject,
-        name: `${cropCycleObject.crop.name} - ${cropCycleObject.season} - ${cropCycleObject.year}`,
-      }));
-
-
-    
-      const initializeInspectionVal = () => {
-        var inspVals = [];
-        inspectionParams.forEach((param) => {
-          inspVals.push({ code: param.code, name: param.displayStr, val: -1 })
-        });
-        setInspectionVal(inspVals);
-      };
-    
-      useEffect(initializeInspectionVal, [inspectionParams]);
-      useEffect(() => {
-        getAdhocActivities(cropCycleId, inspectionVal)
-          .then(setAdHocActivityList)
-          .catch((e) => {
-            console.log("Fetching ad hoc activities failed", e);
-          })
-      }, [inspectionVal, cropCycleId])
     
       function enableAddAdHocActivities() {
         console.log("inside enable adhoc add");
@@ -221,52 +99,26 @@ import {
         console.log(showFollowUpActivity);
         console.log(adHocActivityList);
       }
-      useEffect(enableAddAdHocActivities, [adHocActivityList]);
-
-    const handleChange = ({ target }) => {
-      const { name, value } = target;
-      if (name === 'location') {
-        setLocationCode(value);
-        fetchFieldsForLocation(value);
-        // need to clear the following dropdowns and hence the process cards
-        setFieldId('');
-        setCropCycleId('');
-      } else if (name === 'field') {
-        setFieldId(value);
-        fetchCropCyclesForField(value);
-        // need to clear the following dropdowns and hence the process cards
-        setCropCycleId('');
-      } else if (name === 'crop-cycle') {
-        setCropCycleId(value);
-      }
-    };
     
     const handleInspectionDateChange = (d) => {
         setSelectedInspectionDate(d);
         console.log(d);
       };
-      const handleInspectionChange = ({ target }) => {
-        const { name, value } = target;
-        //target.value = parseInt(value);
-        //console.log(name+":"+value);
-        //console.log(inspectionVal);
-        let inspVals = inspectionVal;
-        inspVals.find(inspV => inspV.name === name).val = parseInt(value);
-        setInspectionVal([...inspVals]);
-        //console.log(inspectionVal);
     
-      }
-      const verifyInspData = () => {
+      //return no of submitted inspection parameters
+      const countInspData = () => {
+        let cnt = 0;
         let indx = 0;
         while (indx < inspectionVal.length) {
-          if (inspectionVal[indx].val === -1) {
-            return 0;
+          if (inspectionVal[indx].val !== -1) {
+            cnt = cnt +1;
           }
           indx++;
         }
-        return 1;
+        return cnt;
       }
     
+      //checks if duplicate Ad hoc actvities are not submitted
       const isDuplicateAdHocEntry = () => {
         console.log("Inside isduplicateAdHocEntry");
         let i = 0;
@@ -293,13 +145,14 @@ import {
       const handleAssigneeChange = () => { };
     
     
-      function processAllInspectionPosting() {
+      function processAllInspectionPosting(cntInspParam) {
         let error = "";
         let info = "Inspection Info successfullly posted";
         
         let inspPostedCnt = 0;
         inspectionVal.forEach((iParam) => {
-         
+         if(iParam.val !== -1)
+         {
           const payload = {
             "inspectionType": iParam.code,
             "resultPositive":Boolean( iParam.val),
@@ -313,7 +166,7 @@ import {
               console.log("Inspection posted for " + iParam.name + " : " + response);
               inspPostedCnt = inspPostedCnt + 1;
     
-              if (inspPostedCnt === inspectionVal.length) {
+              if (inspPostedCnt === cntInspParam) {
                 if (showFollowUpActivity === false) {
                   showAlert(info, 'info');
                   console.log("Inspection count" + inspPostedCnt);
@@ -330,6 +183,9 @@ import {
               showAlert(error, 'error');
     
             });
+         }
+         
+          
         });
       }
     
@@ -371,13 +227,13 @@ import {
     
       }
       const handleSubmit = (e) => {
-    
-        if (verifyInspData() === 0) {
-          showAlert("All inspection parametes need to be filled", 'error');
+        let cntInspParam = countInspData();
+        if (cntInspParam === 0) {
+          showAlert("At least one inspection parameter need to be submitted", 'error');
         } else if (isDuplicateAdHocEntry() === 1) {
           showAlert(" Same activity cannot be added be twice", 'error');
         } else {
-          processAllInspectionPosting();
+          processAllInspectionPosting(cntInspParam);
           console.log("ShowfollowupActivity:" + showFollowUpActivity);
           if (showFollowUpActivity === true) {
             console.log("Inside postAdhoc");
@@ -405,7 +261,7 @@ import {
     };
   
     /**
-     * Returns process cards corresponding to each non-completed processes
+     * Returns Inspection Form containg the different inspection parameters and add hoc activities
      * for the selected crop-cycle
      */
     const getInspectionForm = () => {
@@ -432,64 +288,24 @@ import {
             this crop cycle?'{' '}
           </Typography>
         );
-  
-      // Happy case - return process cards corresponding to each non-completed process
+
+      
       return (
         
          <div>
-            {inspectionParams.map((iParam, index) => {
-                return (
+            <InspectionOptions inspectionVal ={inspectionVal}
+                               setInspectionVal={setInspectionVal}
+                                classStyleObj={classes}
+                                failureHandler={processUpdateFailureHandler}/>
 
-                  <Grid container
-                  direction="row"
-                  justify="center"
-                  alignItems="center"
-                >
-                    <FormControl className={classes.formControl} row>
-                      <FormLabel id='ac-label' className={classes.radioLabel}>{iParam.displayStr}</FormLabel>
-                      <RadioGroup
-                        id={iParam.displayStr}
-                        name={iParam.displayStr}
-                        onChange={handleInspectionChange}
-                        row >
-                        <FormControlLabel value={'1'} control={<Radio />} label="Yes" />
-                        <FormControlLabel value={'0'} control={<Radio />} label="No" />
-
-                      </RadioGroup>
-                    </FormControl>
-                  </Grid>
-
-                )
-              })}
+               
                {showFollowUpActivity ?
-                  followUpActivities.map((x, i) => {
-                    return (
-                      <Grid item xs={12}>
-                        <FormControl className={classes.formControl}>
-                          <InputLabel id='fId-label'>Select Activity</InputLabel>
-                          <Select
-                            id='adHocActivity'
-                            name='adHocActivity'
-                            value={x.id}
-                            onChange={e => handleInputChange(e, i)}
-
-                          >
-                            {adHocActivityList.map((activity) => {
-                              return (
-                                <MenuItem key={activity.code} value={activity.code}>
-                                  {activity.displayStr}
-                                </MenuItem>
-                              );
-                            })}
-                          </Select>
-                          {followUpActivities.length !== 1 && <button
-                            className="mr10"
-                            onClick={() => handleRemoveClick(i)}>Remove this Activity</button>}
-                          {followUpActivities.length - 1 === i && (i < adHocActivityList.length - 1) && <button onClick={handleAddClick}>Add Another Activity</button>}
-                        </FormControl>
-                      </Grid>
-                    )
-                  }) : null
+                  <InspectionAdHocActivity adHocActivityList={adHocActivityList}
+                                          followUpActivities={followUpActivities}
+                                          setFollowUpActivties={setFollowUpActivties}
+                                          classStyleObj={classes}
+                                          />
+                 : null
                 }
               <Grid item xs={12}>
                   <MuiPickersUtilsProvider utils={DateFnsUtils}>
@@ -525,7 +341,6 @@ import {
                     variant='contained'
                     color='primary'
                     onClick={handleSubmit}
-                    
                   >
                     SUBMIT
                   </Button>
@@ -535,6 +350,25 @@ import {
         
     };
   
+    useEffect(() => {
+      getLocations()
+        .then(setLocations)
+        .catch((e) => {
+          console.log('Fetching location list failed', e);
+          showAlert('Fetching location list failed', 'error');
+        });
+    }, []);
+
+    useEffect(() => {
+      getAdhocActivities(cropCycleId, inspectionVal)
+        .then(setAdHocActivityList)
+        .catch((e) => {
+          console.log("Fetching ad hoc activities failed", e);
+        })
+    }, [inspectionVal, cropCycleId])
+    useEffect(enableAddAdHocActivities, [adHocActivityList]);
+
+
     return (
       <FormGroup className={classes.formGroup}>
         <Grid
@@ -550,69 +384,14 @@ import {
               Update Inspection
             </Typography>
           </Grid>
-  
-          {/* Location selection */}
-          <Grid item xs={12}>
-            <FormControl className={classes.formControl}>
-              <InputLabel id='location-label'>Location</InputLabel>
-              <Select
-                id='location'
-                name='location'
-                value={locationCode}
-                onChange={handleChange}
-              >
-                {locations.map((obj) => {
-                  return (
-                    <MenuItem key={obj.displayStr} value={obj.code}>
-                      {obj.displayStr}
-                    </MenuItem>
-                  );
-                })}
-              </Select>
-            </FormControl>
-          </Grid>
-  
-          {/* Field selection */}
-          <Grid item xs={12}>
-            <FormControl className={classes.formControl}>
-              <InputLabel id='field-label'>Field</InputLabel>
-              <Select
-                id='field'
-                name='field'
-                value={fieldId}
-                onChange={handleChange}
-              >
-                {fields.map((obj) => (
-                  <MenuItem key={obj.identifier} value={obj.id}>
-                    {obj.identifier}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-          </Grid>
-  
-          {/* Crop-cycle selection */}
-          <Grid item xs={12}>
-            <FormControl className={classes.formControl}>
-              <InputLabel id='crop-cycle-label'>Crop cycle</InputLabel>
-              <Select
-                id='crop-cycle'
-                name='crop-cycle'
-                value={cropCycleId}
-                onChange={handleChange}
-              >
-                {cropCycles.map((obj) => {
-                  return (
-                    <MenuItem key={obj.name} value={obj.id}>
-                      {obj.name}
-                    </MenuItem>
-                  );
-                })}
-              </Select>
-            </FormControl>
-          </Grid>
-  
-          {/* Process cards for each non-completed processes */}
+          <CropCycleComponent classStyleObj={classes}
+                              locations={locations}
+                              cropCycleId={cropCycleId}
+                              setCropCycleIdHandler={setCropCycleId}
+                              cropCycles={cropCycles}
+                              setCropCyclesHandler={setCropCycles}
+                              failureHandler={processUpdateFailureHandler}/>
+          {/* InspectionForm containing inspection paramters and ad hoc activities if required */}
           <Grid item xs={12}>
             {getInspectionForm()}
           </Grid>
