+++ src/dataclients/InspectionClient.js
@@ -15,18 +15,18 @@ export const getInspectionParams = async () => {
  * API to fetch Adhoc activities based on Inspection Types
  * 
  */
-export const getAdhocActivities = async (fieldCropCycleId,inspectionValues) => {
+export const getAdhocActivities = async (fieldCropCycleId, inspectionValues) => {
   let qParam = "";
-  inspectionValues.forEach((iValue) =>{
-    if(iValue.val === 0)
-    {
+  inspectionValues.forEach((iValue) => {
+    if (iValue.val === 0) {
       qParam = qParam + iValue.code + "=false&"
     }
-    if(iValue.val === 1)
-    {
+    if (iValue.val === 1) {
       qParam = qParam + iValue.code + "=true&"
     }
   })
+  if (fieldCropCycleId === '' || qParam === "")
+    return [];
   return fetch(API_URL + 'fieldCropCycles/' + fieldCropCycleId + '/processes?' + qParam)
     .then(handleErrors)
     .then((response) => response.json());
@@ -37,7 +37,7 @@ export const getAdhocActivities = async (fieldCropCycleId,inspectionValues) => {
  * @param {*} payload 
  */
 
-export const postInspection = async (payload,fieldCropCycleId) => {
+export const postInspection = async (payload, fieldCropCycleId) => {
   return fetch(API_URL + 'fieldCropCycles/' + fieldCropCycleId + '/inspections', {
     method: 'POST',
     body: JSON.stringify(payload),
